% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_cor.R
\name{check_correlation}
\alias{check_correlation}
\title{Check That a Correlation Value Is Within Valid Bounds}
\usage{
check_correlation(cor)
}
\arguments{
\item{cor}{A numeric correlation value.}
}
\value{
Invisible NULL. Throws an error if \code{cor} is outside the range.
}
\description{
Check That a Correlation Value Is Within Valid Bounds
}
\keyword{internal}
