% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaPCoA.R
\name{is_symmetric}
\alias{is_symmetric}
\title{Check Matrix Symmetry}
\usage{
is_symmetric(mat)
}
\arguments{
\item{mat}{A numeric matrix.}
}
\value{
Logical. \code{TRUE} if the matrix is symmetric.
}
\description{
Check Matrix Symmetry
}
\keyword{internal}
